all pagination feeds into same view
