+++ explorer/client/src/components/pagination/Pagination.module.css
@@ -1,52 +1,34 @@
-.pagination {
-    @apply flex flex-row items-center justify-start list-none w-full;
+button.btncontainer,
+button.pagenumber,
+button.nointeract {
+    @apply w-[7vw] sm:w-[32px] h-[32px] px-0 py-2 mx-[1vw] sm:mx-[4px] inline-block
+    rounded border-[1px] border-solid border-sui-grey-50
+    font-sans items-center 
+    text-sui-grey-75 stroke-sui-grey-75 bg-inherit;
 }
 
-.pagination ul {
-    @apply flex flex-row items-center justify-center list-none  p-0 m-0 w-full md:w-min;
+button.btncontainer {
+    @apply hover:bg-sui hover:text-white hover:stroke-white hover:border-sui cursor-pointer;
 }
 
-.arrow {
-    @apply block opacity-50 cursor-default !important;
+button.pagenumber {
+    @apply select-none border-solid border-sui-dark text-sui-dark;
+
+    background: linear-gradient(0deg, #f4fbff, #f4fbff), #fff;
 }
 
-.arrow button {
-    @apply border-transparent cursor-default;
+button.gone {
+    @apply opacity-30;
 }
 
-.pagination ul li {
-    @apply mr-1 hidden md:block;
+.footer {
+    @apply sm:grid sm:grid-cols-12;
 }
 
-.pagination ul li button {
-    @apply text-gray-600 hover:text-gray-800 focus:text-gray-800 font-medium py-2 px-4 cursor-pointer 
-     pr-3  pl-3  border-[1px] rounded-[6px] border-white bg-[#FFFFFF] hover:bg-[#F4FBFF];
-
-    border: 1px solid #e9eaeb;
-    transition: all 0.4s ease;
+.footer > div:first-child {
+    @apply col-span-8 self-center;
 }
 
-button.activepag {
-    @apply bg-[#F4FBFF] text-sui-dark border-sui-dark !important;
-}
-
-.activearrow {
-    @apply block opacity-100 cursor-pointer !important;
-}
-
-.paginationleft svg {
-    transform: rotate(180deg);
-}
-
-.paginationdot {
-    @apply px-1;
-}
-
-.pagilink,
-.paginationdot {
-    @apply block !important;
-}
-
-.paginationdot:hover {
-    @apply bg-[#FFFFFF] hover:bg-[#FFFFFF];
+.footer > div:nth-child(2) {
+    @apply col-span-4;
 }
