--- explorer/client/src/components/pagination/PaginationWrapper.module.css
@@ -1,34 +1,0 @@
-button.btncontainer,
-button.pagenumber,
-button.nointeract {
-    @apply w-[7vw] sm:w-[32px] h-[32px] px-0 py-2 mx-[1vw] sm:mx-[4px] inline-block
-    rounded border-[1px] border-solid border-sui-grey-50
-    font-sans items-center 
-    text-sui-grey-75 stroke-sui-grey-75 bg-inherit;
-}
-
-button.btncontainer {
-    @apply hover:bg-sui hover:text-white hover:stroke-white hover:border-sui cursor-pointer;
-}
-
-button.pagenumber {
-    @apply select-none border-solid border-sui-dark text-sui-dark;
-
-    background: linear-gradient(0deg, #f4fbff, #f4fbff), #fff;
-}
-
-button.gone {
-    @apply opacity-30;
-}
-
-.footer {
-    @apply sm:grid sm:grid-cols-12;
-}
-
-.footer > div:first-child {
-    @apply col-span-8 self-center;
-}
-
-.footer > div:nth-child(2) {
-    @apply col-span-4;
-}
